+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/deneb/ExecutionPayloadSchemaDeneb.java
@@ -54,6 +54,8 @@ import tech.pegasys.teku.spec.datastructures.execution.Transaction;
 import tech.pegasys.teku.spec.datastructures.execution.TransactionSchema;
 import tech.pegasys.teku.spec.datastructures.execution.versions.capella.Withdrawal;
 import tech.pegasys.teku.spec.datastructures.execution.versions.capella.WithdrawalSchema;
+import tech.pegasys.teku.spec.datastructures.execution.versions.electra.DepositReceiptSchema;
+import tech.pegasys.teku.spec.datastructures.execution.versions.electra.ExecutionLayerExitSchema;
 
 public class ExecutionPayloadSchemaDeneb
     extends ContainerSchema17<
@@ -125,6 +127,16 @@ public class ExecutionPayloadSchemaDeneb
   @Override
   public WithdrawalSchema getWithdrawalSchemaRequired() {
     return getWithdrawalSchema();
+  }
+
+  @Override
+  public DepositReceiptSchema getDepositReceiptSchemaRequired() {
+    throw new IllegalStateException("Attempted to get a deposit receipt schema from deneb");
+  }
+
+  @Override
+  public ExecutionLayerExitSchema getExecutionLayerExitSchemaRequired() {
+    throw new IllegalStateException("Attempted to get a exit schema from bellatrix");
   }
 
   public WithdrawalSchema getWithdrawalSchema() {