+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/electra/ExecutionPayloadElectraImpl.java
@@ -22,8 +22,8 @@ import tech.pegasys.teku.infrastructure.bytes.Bytes20;
 import tech.pegasys.teku.infrastructure.ssz.SszList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteVector;
-import tech.pegasys.teku.infrastructure.ssz.containers.Container17;
-import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema17;
+import tech.pegasys.teku.infrastructure.ssz.containers.Container19;
+import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema19;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt256;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
@@ -33,7 +33,7 @@ import tech.pegasys.teku.spec.datastructures.execution.Transaction;
 import tech.pegasys.teku.spec.datastructures.execution.versions.capella.Withdrawal;
 
 public class ExecutionPayloadElectraImpl
-    extends Container17<
+    extends Container19<
         ExecutionPayloadElectraImpl,
         SszBytes32,
         SszByteVector,
@@ -51,11 +51,13 @@ public class ExecutionPayloadElectraImpl
         SszList<Transaction>,
         SszList<Withdrawal>,
         SszUInt64,
-        SszUInt64>
+        SszUInt64,
+        SszList<DepositReceipt>,
+        SszList<ExecutionLayerExit>>
     implements ExecutionPayloadElectra {
 
   public ExecutionPayloadElectraImpl(
-      ContainerSchema17<
+      ContainerSchema19<
               ExecutionPayloadElectraImpl,
               SszBytes32,
               SszByteVector,
@@ -73,7 +75,9 @@ public class ExecutionPayloadElectraImpl
               SszList<Transaction>,
               SszList<Withdrawal>,
               SszUInt64,
-              SszUInt64>
+              SszUInt64,
+              SszList<DepositReceipt>,
+              SszList<ExecutionLayerExit>>
           schema,
       TreeNode backingNode) {
     super(schema, backingNode);
@@ -97,7 +101,9 @@ public class ExecutionPayloadElectraImpl
       SszList<Transaction> transactions,
       SszList<Withdrawal> withdrawals,
       SszUInt64 blobGasUsed,
-      SszUInt64 excessBlobGas) {
+      SszUInt64 excessBlobGas,
+      SszList<DepositReceipt> depositReceipts,
+      SszList<ExecutionLayerExit> exits) {
     super(
         schema,
         parentHash,
@@ -116,7 +122,9 @@ public class ExecutionPayloadElectraImpl
         transactions,
         withdrawals,
         blobGasUsed,
-        excessBlobGas);
+        excessBlobGas,
+        depositReceipts,
+        exits);
   }
 
   @Override
@@ -225,7 +233,21 @@ public class ExecutionPayloadElectraImpl
   }
 
   @Override
+  public SszList<DepositReceipt> getDepositReceipts() {
+    return getField17();
+  }
+
+  @Override
+  public SszList<ExecutionLayerExit> getExits() {
+    return getField18();
+  }
+
+  @Override
   public List<TreeNode> getUnblindedTreeNodes() {
-    return List.of(getTransactions().getBackingNode(), getWithdrawals().getBackingNode());
+    return List.of(
+        getTransactions().getBackingNode(),
+        getWithdrawals().getBackingNode(),
+        getDepositReceipts().getBackingNode(),
+        getExits().getBackingNode());
   }
 }