+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/electra/ExecutionPayloadHeaderSchemaElectra.java
@@ -17,7 +17,9 @@ import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFi
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.BLOB_GAS_USED;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.BLOCK_HASH;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.BLOCK_NUMBER;
+import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.DEPOSIT_RECEIPTS_ROOT;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.EXCESS_BLOB_GAS;
+import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.EXITS_ROOT;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.EXTRA_DATA;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.FEE_RECIPIENT;
 import static tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadFields.GAS_LIMIT;
@@ -36,7 +38,7 @@ import java.util.function.Consumer;
 import tech.pegasys.teku.infrastructure.bytes.Bytes20;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteList;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszByteVector;
-import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema17;
+import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema19;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt256;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
@@ -51,7 +53,7 @@ import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeaderBui
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeaderSchema;
 
 public class ExecutionPayloadHeaderSchemaElectra
-    extends ContainerSchema17<
+    extends ContainerSchema19<
         ExecutionPayloadHeaderElectraImpl,
         SszBytes32,
         SszByteVector,
@@ -69,7 +71,9 @@ public class ExecutionPayloadHeaderSchemaElectra
         SszBytes32,
         SszBytes32,
         SszUInt64,
-        SszUInt64>
+        SszUInt64,
+        SszBytes32,
+        SszBytes32>
     implements ExecutionPayloadHeaderSchema<ExecutionPayloadHeaderElectraImpl> {
 
   private final ExecutionPayloadHeaderElectraImpl defaultExecutionPayloadHeader;
@@ -94,7 +98,9 @@ public class ExecutionPayloadHeaderSchemaElectra
         namedSchema(TRANSACTIONS_ROOT, SszPrimitiveSchemas.BYTES32_SCHEMA),
         namedSchema(WITHDRAWALS_ROOT, SszPrimitiveSchemas.BYTES32_SCHEMA),
         namedSchema(BLOB_GAS_USED, SszPrimitiveSchemas.UINT64_SCHEMA),
-        namedSchema(EXCESS_BLOB_GAS, SszPrimitiveSchemas.UINT64_SCHEMA));
+        namedSchema(EXCESS_BLOB_GAS, SszPrimitiveSchemas.UINT64_SCHEMA),
+        namedSchema(DEPOSIT_RECEIPTS_ROOT, SszPrimitiveSchemas.BYTES32_SCHEMA),
+        namedSchema(EXITS_ROOT, SszPrimitiveSchemas.BYTES32_SCHEMA));
 
     final ExecutionPayloadElectraImpl defaultExecutionPayload =
         new ExecutionPayloadSchemaElectra(specConfig).getDefault();
@@ -113,7 +119,9 @@ public class ExecutionPayloadHeaderSchemaElectra
   public LongList getBlindedNodeGeneralizedIndices() {
     return LongList.of(
         getChildGeneralizedIndex(getFieldIndex(TRANSACTIONS_ROOT)),
-        getChildGeneralizedIndex(getFieldIndex(WITHDRAWALS_ROOT)));
+        getChildGeneralizedIndex(getFieldIndex(WITHDRAWALS_ROOT)),
+        getChildGeneralizedIndex(getFieldIndex(DEPOSIT_RECEIPTS_ROOT)),
+        getChildGeneralizedIndex(getFieldIndex(EXITS_ROOT)));
   }
 
   @Override
@@ -162,6 +170,8 @@ public class ExecutionPayloadHeaderSchemaElectra
         SszBytes32.of(executionPayload.getTransactions().hashTreeRoot()),
         SszBytes32.of(executionPayload.getWithdrawals().hashTreeRoot()),
         SszUInt64.of(executionPayload.getBlobGasUsed()),
-        SszUInt64.of(executionPayload.getExcessBlobGas()));
+        SszUInt64.of(executionPayload.getExcessBlobGas()),
+        SszBytes32.of(executionPayload.getDepositReceipts().hashTreeRoot()),
+        SszBytes32.of(executionPayload.getExits().hashTreeRoot()));
   }
 }